refactor(calendar): refactor event preview card into its own component
