+++ src/app/calendar-app/calendar-event-card.component.html
@@ -1,0 +1,32 @@
+<mat-card class="upcomingEventCard">
+    <mat-card-header>
+        <mat-card-title>
+            {{ e.title }}
+            <span *ngIf="e.ongoing" style="font-weight: bold;">
+                ONGOING
+            </span>
+        </mat-card-title>
+        <mat-card-subtitle>
+            <span *ngIf="!e.ongoing; else started"> Starts </span>
+            <ng-template #started> Started </ng-template>
+            {{ e.dtstart.calendar() }}
+        </mat-card-subtitle>
+        <mat-card-subtitle *ngIf="e.dtend">
+            Ends {{ e.dtend.calendar() }}
+        </mat-card-subtitle>
+        <mat-card-subtitle *ngIf="e.recurringFrequency">
+            {{ e.recurringFrequency }} event
+        </mat-card-subtitle>
+        <mat-card-subtitle *ngIf="e.location">
+            Location: {{ e.location }}
+        </mat-card-subtitle>
+    </mat-card-header>
+    <mat-card-content>
+        <p>
+            {{ e.description }}
+        </p>
+    </mat-card-content>
+    <mat-card-actions>
+        <button mat-button *ngIf="editable" (click)="edit.emit()"> Edit </button>
+    </mat-card-actions>
+</mat-card>
