+++ src/app/calendar-app/calendar-overview.component.html
@@ -1,40 +1,14 @@
 <div style="padding-left: 25px; padding-right: 25px;">
     <h2> Upcoming events </h2>
 
-    <mat-card class="upcomingEventCard"
+    <app-calendar-event-card
         *ngFor="let e of visible_events.slice(0, event_limit)"
+        [editable]="true"
+        [event]="e"
+        (edit)="editEvent.emit(e)"
+
     >
-        <mat-card-header>
-            <mat-card-title>
-                {{ e.title }}
-                <span *ngIf="e.ongoing" style="font-weight: bold;">
-                    ONGOING
-                </span>
-            </mat-card-title>
-            <mat-card-subtitle>
-                <span *ngIf="!e.ongoing; else started"> Starts </span>
-                <ng-template #started> Started </ng-template>
-                {{ e.dtstart.calendar() }}
-            </mat-card-subtitle>
-            <mat-card-subtitle *ngIf="e.dtend">
-                Ends {{ e.dtend.calendar() }}
-            </mat-card-subtitle>
-            <mat-card-subtitle *ngIf="e.recurringFrequency">
-                {{ e.recurringFrequency }} event
-            </mat-card-subtitle>
-            <mat-card-subtitle *ngIf="e.location">
-                Location: {{ e.location }}
-            </mat-card-subtitle>
-        </mat-card-header>
-        <mat-card-content>
-            <p>
-                {{ e.description }}
-            </p>
-        </mat-card-content>
-        <mat-card-actions>
-            <button mat-button (click)="editEvent.emit(e.event)"> Edit </button>
-        </mat-card-actions>
-    </mat-card>
+    </app-calendar-event-card>
 
     <div *ngIf="visible_events.length - event_limit; let remaining">
         <mat-card *ngIf="remaining > 0" class="upcomingEventCard">
